refactor(input): exported to it's own header
